third_party/WebKit/LayoutTests/http/tests/background_sync/interfaces.html

<!doctype html>
<html>
  <head>
    <title>Background Sync API: Verifies that the right Background Sync API interfaces get exposed.</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="../serviceworker/resources/test-helpers.js"></script>
  </head>
  <body>
    <script>
      // Tests that the expected Background Sync API interfaces are being exposed on both
      // the Service Worker and Document global scopes.
      service_worker_test(
          'resources/interfaces-worker.js',
          'Exposure of interfaces in a Service Worker.');

      test(function() {
          assert_own_property(self, 'SyncManager', 'SyncManager needs to be exposed as a global.');

          assert_own_property(SyncManager.prototype, 'register');
          assert_own_property(SyncManager.prototype, 'getRegistration');
          assert_own_property(SyncManager.prototype, 'getRegistrations');
          assert_own_property(SyncManager.prototype, 'permissionState');

      }, 'SyncManager should be exposed and have the expected interface in a Document.');

      test(function() {
          assert_own_property(self, 'SyncRegistration', 'SyncRegistration needs to be exposed as a global.');

          // FIXME: Assert existence of the attributes when they are properly
          // exposed in the prototype chain. https://crbug.com/43394

          assert_own_property(SyncRegistration.prototype, 'finished');
          assert_own_property(SyncRegistration.prototype, 'unregister');
          assert_own_property(SyncRegistration.prototype, 'tag');

      }, 'SyncRegistration should be exposed and have the expected interface in a Document.');
    </script>
  </body>
</html>
